moved queries and routes to separate folders, added app.js, added twilio.js
